Initial drop
This is 1/2 original firmware - with my minor changes for debugging - and a non-functioning but skeleton server instance.  The server in particular needs some pretty drastic refactoring if I'm going to support all the things that I would really like to support.  A single file for the server is a pure reflection on my desire to get a simple hello world to work.  Having said that there is support here for multiple boards which is outside of my use case.

I will defer iterating on the iBoardBot firmware at this time.  While I think that there is the potential for improvements here I will be focusing on the server and supporting the 'stock' firmware minus changes to the URL.