Better debug on firmware; good erase
Added the erase functionality and, in order to figure out how erase was supposed to work on the firmware, added a bit of information to the debugging data set to serial from the firmware.

There appear to be continuing issues with the firmware.  I've seen so far:
  1/  Firmware decide that it doesn't know what network to connect to
  2/  Firmware doesn't send any string in the ID_IWBB field

All of these are likely due to the usage of the AT command set in order to communicate with the ESP8266 from the Arduino.  This entire approach has a number of huge disadvantages (the AT command set and interpretation of string results for example) and also exhibits all the major weaknesses in the hardware architecture (Arduino as a microcontroller is far less powerful than the ESP8266).