Added forecast data
You have to call the OpenWeather 'forecast' API in order to get day ranges.  The min and max values returned for a single 'full weather' forecast are presumably some min and max at that time within a range or area covered by the city.  Forecast API solves this problem by giving you many data points across a day.  It is important to maintain a day low and high though so that it doesn't increase or decrease in the wrong direction as the day gets shorter and you have less data points about the range of values.