Simplified weather scheduling, added filled text
Woke up this morning and realized that I really wanted the weather to be updated at specific hours of the day and not using a periodicity.  After I made that change I was also unhappy with how hard it was to read the time and temperature from across the kitchen so I added a way to display filled in text.  Much darker now and easier to read.