Filled text implementation changed
The prior implementation of filled text caused a ton of redundant lines to be drawn given the thickness of the pen.  This changes that to better capture the shape but with less lines.  Not 100% happy with the overall approach here as the transformation to lines via OpenCV has some surprising results.  Perhaps it is time to write my own raster -> lines routine with a simple scanline approach.