Support for an incremental weather display
I love the weather display but the complete erase and then redraw was starting to be messy and I wish that I could see a summary at the end of the day for how the day went.

This change introduces an incremental mode (and gets rid of the old 'update the weather display' where I would erase and then redraw only a portion of it -- I never used it) that solves both problems.