Stop long lines at the end of drawing
I changed the SERVO2_LIFT from 890 to 1110 because it was rotating further during warm weather and as a result rotating far enough that it left a line from the last point to draw back to the home position.  If the last position drawn was close enough to home then the servo wouldn't rotate far enough in order to leave a mark.  But really this entire firmware is too simplistic.  It doesn't wait for things to happen and relies on timing and magic in order to ensure that things happen on time.  5 years in to having this with various fixes to ensure that it continues to run and the cracks are showing.